apps: load_fw: Separate rproc operations from main app
Separate rproc operations from the same file as where main
application is.
